+++ pymidil/transports/redis/__init__.py
@@ -1,0 +1,3 @@
+from pymidil.transports.redis.producer import RedisProducer, RedisProducerEventConfig
+
+__all__ = ["RedisProducer", "RedisProducerEventConfig"]